Moved auth stuff into users
