+++ server/users/userRoutes.js
@@ -1,0 +1,6 @@
+var AuthController = require('./authController');
+
+module.exports = function(route) {
+  route.post('/login', AuthController.userLogin);
+  route.post('/signup', AuthController.userSignup);
+};
